Add parameter to temporal functions
