+++ .changeset/breezy-hotels-call.md
@@ -2,4 +2,4 @@
 "@neo4j/cypher-builder": patch
 ---
 
-Add parameters to date function
+Add timezone parameter to temporal functions
