+++ packages/cypher-builder/src/Cypher.ts
@@ -77,18 +77,13 @@ export {
 export { plus, minus } from "./expressions/operations/math";
 
 // --Functions
-export { CypherFunction as Function } from "./expressions/functions/CypherFunction";
+export { CypherFunction as Function } from "./expressions/functions/CypherFunctions";
 
 export {
     coalesce,
     point,
     distance,
     pointDistance,
-    cypherDatetime as datetime,
-    cypherDate as date,
-    cypherLocalTime as localtime,
-    cypherLocalDatetime as localdatetime,
-    cypherTime as time,
     labels,
     count,
     min,
@@ -96,7 +91,15 @@ export {
     avg,
     sum,
     randomUUID,
-} from "./expressions/functions/CypherFunction";
+} from "./expressions/functions/CypherFunctions";
+
+export {
+    cypherDatetime as datetime,
+    cypherDate as date,
+    cypherLocalTime as localtime,
+    cypherLocalDatetime as localdatetime,
+    cypherTime as time,
+} from "./expressions/functions/TemporalFunctions";
 
 export * from "./expressions/functions/StringFunctions";
