+++ packages/cypher-builder/src/expressions/functions/CypherFunctions.ts
@@ -102,85 +102,6 @@ export function labels(nodeRef: Variable): CypherFunction {
 }
 
 /**
- * @see [Cypher Documentation](https://neo4j.com/docs/cypher-manual/current/functions/temporal/#functions-datetime)
- * @group Expressions
- * @category Cypher Functions
- */
-export function cypherDatetime(): CypherFunction {
-    return new CypherFunction("datetime");
-}
-
-/**
- * @see [Cypher Documentation](https://neo4j.com/docs/cypher-manual/current/functions/temporal/#functions-date)
- * @group Expressions
- * @category Cypher Functions
- * @example
- *
- * Using date without parameters:
- *
- * ```ts
- * Cypher.Date()
- * ```
- *
- * _Cypher:_
- * ```cypher
- * date()
- * ```
- *
- * ---
- *
- * Date with parameters:
- *
- * ```ts
- * Cypher.Date(new Cypher.param('9999-01-01'))
- * ```
- *
- * _Cypher:_
- * ```cypher
- * date($param1)
- * ```
- *
- * _Params:_
- * ```json
- * {
- *   param1: "9999-01-01"
- * }
- * ```
- *
- *
- */
-export function cypherDate(timezone?: Expr): CypherFunction {
-    return new CypherFunction("date", timezone ? [timezone] : undefined);
-}
-
-/**
- * @see [Cypher Documentation](https://neo4j.com/docs/cypher-manual/current/functions/scalar/#functions-coalesce)
- * @group Expressions
- * @category Cypher Functions
- */
-export function cypherLocalDatetime(): CypherFunction {
-    return new CypherFunction("localdatetime");
-}
-
-/**
- * @see [Cypher Documentation](https://neo4j.com/docs/cypher-manual/current/functions/temporal/#functions-localdatetime)
- * @group Expressions
- * @category Cypher Functions
- */
-export function cypherLocalTime(): CypherFunction {
-    return new CypherFunction("localtime");
-}
-
-/**
- * @see [Cypher Documentation](https://neo4j.com/docs/cypher-manual/current/functions/temporal/#functions-time)
- * @group Expressions
- * @category Cypher Functions
- */
-export function cypherTime(): CypherFunction {
-    return new CypherFunction("time");
-}
-
-/**
  * @see [Cypher Documentation](https://neo4j.com/docs/cypher-manual/current/functions/aggregating/#functions-count)
  * @group Expressions
  * @category Cypher Functions
