+++ packages/cypher-builder/src/expressions/functions/ListFunctions.ts
@@ -20,7 +20,7 @@
 import type { Variable } from "../..";
 import type { CypherEnvironment } from "../../Environment";
 import type { Expr } from "../../types";
-import { CypherFunction } from "./CypherFunction";
+import { CypherFunction } from "./CypherFunctions";
 
 /**
  * @see [Cypher Documentation](https://neo4j.com/docs/cypher-manual/current/functions/scalar/#functions-size)
